+++ plugins/woocommerce-admin/packages/components/src/search/index.js
@@ -14,7 +14,7 @@ import classnames from 'classnames';
  * Internal dependencies
  */
 import Autocomplete from './autocomplete';
-import { coupons, customers, product, productCategory, taxes, variations } from './autocompleters';
+import { countries, coupons, customers, product, productCategory, taxes, variations } from './autocompleters';
 import Tag from '../tag';
 
 /**
@@ -66,18 +66,20 @@ class Search extends Component {
 
 	getAutocompleter() {
 		switch ( this.props.type ) {
+			case 'countries':
+				return countries;
+			case 'coupons':
+				return coupons;
+			case 'customers':
+				return customers;
 			case 'products':
 				return product;
 			case 'product_cats':
 				return productCategory;
-			case 'coupons':
-				return coupons;
 			case 'taxes':
 				return taxes;
 			case 'variations':
 				return variations;
-			case 'customers':
-				return customers;
 			default:
 				return {};
 		}
@@ -208,11 +210,12 @@ Search.propTypes = {
 	 * The object type to be used in searching.
 	 */
 	type: PropTypes.oneOf( [
+		'countries',
+		'coupons',
+		'customers',
+		'orders',
 		'products',
 		'product_cats',
-		'orders',
-		'customers',
-		'coupons',
 		'taxes',
 		'variations',
 	] ).isRequired,
@@ -225,7 +228,10 @@ Search.propTypes = {
 	 */
 	selected: PropTypes.arrayOf(
 		PropTypes.shape( {
-			id: PropTypes.number.isRequired,
+			id: PropTypes.oneOfType( [
+				PropTypes.number,
+				PropTypes.string,
+			] ).isRequired,
 			label: PropTypes.string.isRequired,
 		} )
 	),